refactor: remove redundant overloads and improve readability

+++ src/mkpsxiso/iso.h
@@ -74,7 +74,7 @@ namespace iso
 		DirTreeClass* parent; // Non-owning
 		
 		/// Internal function for generating and writing directory records
-		bool WriteDirEntries(cd::IsoWriter* writer, const DIRENTRY& dir, const DIRENTRY& parentDir, const unsigned short totalDirs) const;
+		bool WriteDirEntries(cd::IsoWriter* writer, const DIRENTRY& dir, const DIRENTRY& parentDir, const int totalDirs) const;
 
 		/// Internal function for recursive path table generation
 		std::unique_ptr<PathTableClass> GenPathTableSub(unsigned short& index, unsigned short parentIndex) const;
@@ -161,7 +161,7 @@ namespace iso
 		 *	root		   - Root directory
 		 *  totalDirs	   - Total number of directories. Only usefull for games built with the latest sony mastering tool
 		 */
-		bool WriteDirectoryRecords(cd::IsoWriter* writer, const DIRENTRY& root, unsigned short totalDirs);
+		bool WriteDirectoryRecords(cd::IsoWriter* writer, const DIRENTRY& root, int totalDirs);
 
 		void SortDirectoryEntries(const bool byOrder, const bool byLBA = false);
